Move all code related to resource types and principals to dedicated modules
