+++ mcmd/commands/disable.py
@@ -1,6 +1,7 @@
 import mcmd.config.config as config
 from mcmd import io
-from mcmd.client.molgenis_client import ResourceType, post, ensure_resource_exists
+from mcmd.client.molgenis_client import post
+from mcmd.utils.resources import ensure_resource_exists, ResourceType
 from mcmd.command import command
 from mcmd.commands._registry import arguments
 from mcmd.io import highlight
